Testcase for new messageHistory for both Datasetlifecycle and Job added
