+++ common/models/job.json
@@ -73,7 +73,7 @@
         "messageHistory": {
             "type": "embedsMany",
             "model": "Message",
-            "property": "messageHistory",
+            "property": "MessageHistory",
             "options": {
                 "validate": false,
                 "forceId": false,
