+++ common/models/message.json
@@ -1,7 +1,7 @@
 {
     "name": "Message",
     "description": "Embedded information used inside DatasetLifecycle or Job to keep track of history.",
-    "base": "PersistedModel",
+    "base": "Model",
     "idInjection": true,
     "strict": true,
     "options": {
@@ -11,30 +11,27 @@
         "id": {
             "type": "string",
             "id": "true",
-            "required": true,
-            "description": "id currently needed by limitation in embedsmanny"
+            "defaultFn": "uuid",
+            "description": "id currently needed by limitation in embedsmany"
         },
         "shortMessage": {
             "type": "string",
             "required": true,
-            "index": true,
             "description": "Short description of message, like datasetArchiveSuccesful"
         },
         "sender": {
             "type": "string",
             "required": true,
-            "index": true,
-            "description": "User or role who triggered the message"
+            "description": "Email of user or role who triggered the message"
         },
         "when": {
             "type": "Date",
-            "index": true,
             "default": "$now",
             "description": "Time when message was received. Will be filled automatically by the server, format according to chapter 5.6 internet date/time format in RFC 3339."
         },
-        "comment": {
+        "payload": {
             "type": "Object",
-            "description": "Additional information relevant for this message"
+            "description": "Additional information relevant for this message. Please note this is an object"
         }
     },
     "validations": [],
